dipex v0.1.0 Cache

Functions

get(key)

get_and_or_update(key, val)

Specifically built for gpio status

If gpio status has not been set (first boot) and it cannot be found in ETS set it to the given value (nil/“on”/“off”)

If a record is found and the given value is not the same as the stored value update stored value to new value

However if the record is found and the querying value is nil return the response

The idea here is that you can call this with nil to ensure you don’t update the record and change the gpio status

When you call it with “on” or “off” and the record is different than your query of “on” or “off” the record is updated and the new record is returned

get_or_set(key, val)

init()

set(key, data)